Aligned symbols with established terminology
Aligned several symbol names with established terminology in the CIF
field, and somewhat with the terminology used in the CIF 2.0 paper.
Combined a couple of productions for string data values, where the
paper does not use distinguishing terms for them.

Incorporated the three forms of line termination, instead of
hand waving about line terminator normalization.  This introduces
a slight ambiguity, but that can be resolved in eitgher direction
without changing which inputs are accepted by the grammar.