Cif2 dimension notation fix up (#121)
* Changed: updated the description of the 'Dimension' content type based
on the discussion related to the '_type.dimension' data item. Updated
the '_type.dimension' data item description since it was slightly
ambiguous. The old description implied that the value consists of a
text string that is placed withing bounding brackets and not that
the bounding brackets are a part of the text string.

* Fixed: changed the values of the '_type.dimension' data item to conform
to the new notation.

* Fixed: updated the '_type.dimension' data item values to fit the new
convention. Corrected several unrecognised data names
(i.e. '_enumeration_default' -> '_enumeration_default.value').
Corrected several default values to use the correct CIF2 list notation
(list values inside square brackets must be separated by a whitespace
symbol and not a comma).

* Fixed: replacing the '_enumeration_default.value' data names with
'_enumeration.default' since the former were incorrectly added in
the previous commit.

* Fixed: updating the '_type.dimension' values to adhere to the new
syntax in the 'cif_img.dic'.