Improve semantics of `_diffrn.id` pointers
The `CELL_MEASUREMENT` category now has two `_diffrn.id` pointers, one to specify the data collection to which the cell measurement applies, and one to the conditions under which the cell was measured (`.condition_id`). Normally these coincide and the latter is not necessary, but legacy point detector experiments can sometimes involve measuring the cell under differing conditions prior to data collection.