Improve summaries in preparation for release.
As the git repository stores the details of all changes, only the most significant changes need to be recorded in the dictionary audit area. The significant changes are generally new categories and data names, or deprecation of old data names. Corrections or changes to types and units have generally been covered by generic statements.