Improve DICTIONARY_AUDIT summaries before release. (#361)
* Improve summaries in preparation for release.

As the git repository stores the details of all changes, only the most significant changes need to be recorded in the dictionary audit area. The significant changes are generally new categories and data names, or deprecation of old data names. Corrections or changes to types and units have generally been covered by generic statements.

* More general DDL version 4 change note.

The original said the change was to version 4.0.1. The current conformance is version 4.1.0. Rather than note every change of DDL version, the step to version 4 is the most significant.

* Create a detailed changelog for developer reference.

As the released dictionaries contain only a summary of the most important changes, a detailed changelog has been added to preserve the more detailed change lists that build up during the development of a new version.
---------

Co-authored-by: Antanas Vaitkus <antanas.vaitkus90@gmail.com>