Repurpose STRUCTURE category.
This commit properly defines the meaning of a "structure" in
the following steps:

1. All categories describing the space group are assigned a "_space_group.id"
key data name or pointer. This makes explicit the fact that they all relate to
the same space group.

2. The STRUCTURE category now holds a description of structure, which consists
of a pointer to the space group and the diffraction conditions that it
relates to.

3. The ATOM_SITE and CELL categories now include a key data name pointer to
`_structure.id`, identifying which structure an atom or cell relates to.

4. More controversially, `_cell.diffrn_id` has been replaced by
`_structure.diffrn_id`. To strictly follow our rules, it should be
deprecated, but it is unlikely to have been used since the release.