Remove dot key data names (#522)
* Remove the _atom_type.key data item

This item was originally introduced as an arbitrary unique
identifier of the ATOM_TYPE category. However, the item was
never used in practice and at some point the category key was
changed back to _atom_type.symbol. See GH #520 for more details.

* Rename _space_group_generator.key to _space_group_generator.id

The name was change to be more inline with other category key names. This name
was introduced during the migration to DDLM and merging of the CIF_CORE and
CIF_SYM dictionaries, and does not seem to have even been used anywhere.