Changes to modulated structures imports for consistency.
Some definitions have a _definition.update in both template block
and main definition, which causes an import error. The date on
the main definition is preferred, so dates have been removed.

Also, if the presence of _method.expression implies a Derived
value, then some import blocks need a _type.source of Derived.